ecdsa_new_priv_openssh: use correct free function on failure.
Thanks to Ralf Esswein for pointing out the slip, which should have
been corrected as part of 867e69187.

It didn't cause a test failure, because the ecdsa and eddsa structs
are currently so similar in layout (they differ only in the pointed-to
public key point structure, and on this failure path that pointer is
NULL anyway). As a result it's rather hard to add a regression test
against it failing again; I think I just have to chalk this one up to
the lack of OO-aware type checking when doing this kind of manual
vtable management in C, unfortunately.